refactor: move handleClozeDeletions into a own file
